+++ server/src/main/java/com/tuofan/setting/controller/SysMyClassController.java
@@ -4,11 +4,13 @@ package com.tuofan.setting.controller;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.tuofan.core.LoginConstants;
 import com.tuofan.core.Result;
+import com.tuofan.course.service.ICourseMainService;
 import com.tuofan.setting.entity.SysClassNo;
 import com.tuofan.setting.entity.SysClassRoom;
 import com.tuofan.setting.entity.SysMyClass;
 import com.tuofan.setting.service.ISysClassNoService;
 import com.tuofan.setting.service.ISysMyClassService;
+import com.tuofan.student.service.IStudentMainService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -29,6 +31,9 @@ public class SysMyClassController {
 
     @Autowired
     private ISysMyClassService iSysMyClassService;
+
+    @Autowired
+    private ICourseMainService iCourseMainService;
 
 
     @GetMapping("list")
@@ -63,6 +68,9 @@ public class SysMyClassController {
 
     @PostMapping("delete")
     public Result save(@RequestParam Integer id) {
+        QueryWrapper query = new QueryWrapper();
+        query.eq("class_id", id);
+        if (iCourseMainService.list(query).size() > 0) return Result.error("有课程，不能删除");
         iSysMyClassService.removeById(id);
         return Result.ok();
     }
